backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roughcut-backend"
version = "0.1.0"
description = "Local-first AI rough-cut editing backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.28.1",
  "faster-whisper>=1.1.1",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.20",
  "uvicorn[standard]>=0.34.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

